ai-service/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
python-multipart==0.0.9
langgraph==0.2.28
langchain-core==0.3.15
langchain-community==0.3.1
langchain-huggingface==0.1.0
langchain-nvidia-ai-endpoints==0.3.5
sentence-transformers==3.1.1
faiss-cpu==1.8.0
PyMuPDF==1.24.6
rapidfuzz==3.9.1
networkx==3.3
httpx==0.27.0
pydantic==2.8.0
python-dotenv==1.0.1
mcp==1.3.0
aiosqlite==0.20.0
lxml==5.3.0
pytest==8.3.2
pytest-asyncio==0.23.8
